[chore] #64 - SessionStartView+Extension 폴더 삭제 및 코드 이동
